refactor(contacts): update Instagram types and analysis result handling
- Consolidate Instagram-related types into a single import from the instagram module for better organization
- Update analyzeConversation and batchAnalyzeConversations functions to return AnalysisResult instead of GeminiAnalysisResult for consistency
- Adjust type definitions in storeContacts function to align with the new AnalysisResult type
- Improve code clarity by removing redundant type definitions from contacts.ts
