+++ src/types/index.d.ts
@@ -14,3 +14,5 @@ export type SiteConfig = {
     x: string;
   }
 }
+
+export * from './instagram';
